misc: Move Jellyfin Artist iterator to separate file
