+++ backend/mediaprovider/jellyfin/iterators.go
@@ -16,8 +16,6 @@ const (
 	AlbumSortArtistAZ       string = "Artist (A-Z)"
 	AlbumSortYearAscending  string = "Year (ascending)"
 	AlbumSortYearDescending string = "Year (descending)"
-
-	ArtistSortNameAZ string = "Name (A-Z)"
 )
 
 func (j *jellyfinMediaProvider) AlbumSortOrders() []string {
@@ -28,12 +26,6 @@ func (j *jellyfinMediaProvider) AlbumSortOrders() []string {
 		AlbumSortArtistAZ,
 		AlbumSortYearAscending,
 		AlbumSortYearDescending,
-	}
-}
-
-func (j *jellyfinMediaProvider) ArtistSortOrders() []string {
-	return []string{
-		ArtistSortNameAZ,
 	}
 }
 
@@ -122,43 +114,6 @@ func (j *jellyfinMediaProvider) IterateTracks(searchQuery string) mediaprovider.
 		}
 	}
 	return helpers.NewTrackIterator(fetcher, j.prefetchCoverCB)
-}
-
-func (j *jellyfinMediaProvider) IterateArtists(sortOrder string, filter mediaprovider.ArtistFilter) mediaprovider.ArtistIterator {
-	var jfSort jellyfin.Sort
-
-	if sortOrder == "" {
-		sortOrder = ArtistSortNameAZ // default
-	}
-	switch sortOrder {
-	case ArtistSortNameAZ:
-		jfSort.Field = jellyfin.SortByName
-		jfSort.Mode = jellyfin.SortAsc
-	}
-
-	fetcher := func(offs, limit int) ([]*mediaprovider.Artist, error) {
-		ar, err := j.client.GetAlbumArtists(jellyfin.QueryOpts{
-			Sort:   jfSort,
-			Paging: jellyfin.Paging{StartIndex: offs, Limit: limit},
-		})
-		if err != nil {
-			return nil, err
-		}
-		return sharedutil.MapSlice(ar, toArtist), nil
-	}
-
-	return helpers.NewArtistIterator(fetcher, filter, j.prefetchCoverCB)
-}
-
-func (j *jellyfinMediaProvider) SearchArtists(searchQuery string, filter mediaprovider.ArtistFilter) mediaprovider.ArtistIterator {
-	fetcher := func(offs, limit int) ([]*mediaprovider.Artist, error) {
-		sr, err := j.client.Search(searchQuery, jellyfin.TypeArtist, jellyfin.Paging{StartIndex: offs, Limit: limit})
-		if err != nil {
-			return nil, err
-		}
-		return sharedutil.MapSlice(sr.Artists, toArtist), nil
-	}
-	return helpers.NewArtistIterator(fetcher, filter, j.prefetchCoverCB)
 }
 
 // Creates the Jellyfin filter to implement the given mediaprovider filter,
